Pull generic parts of NewsCollectionViewCell out into SideScrollingCollectionViewCell.
Renamed NewsArticle to CellArticle. Can probably do better...
